+++ src/blockly-blocks/blocks.js
@@ -31,5 +31,5 @@ import "./tephra/block-calculate-tephra-vei-wind";
 import "./tephra/block-risk-level";
 import "./seismic/block-seismic-gps-stations";
 import "./seismic/block-seismic-graph-gps-position";
-import "./seismic/block-seismic-strain-rate";
+import "./seismic/block-seismic-deformation-buildup";
 import "./deformation/block-deformation-create-sim";